Swap the endianness of read accelerometer values
When testing on an ADXL345 accelerometer, that shares a device-id with
the ADXL343 and looks mostly register-compatible, the measurements
returned by accel_norm() and accel_raw() do not look right.

The datasheet for both ADXL345 and ADXL343 states, that

> The outputdata is twos complement, with DATA x 0 as the
> least significant byte and DATA x 1 as the most significant byte

e.g. little endian. Change lib.rs to reflect this endianness.